deepreview/pyproject.toml
[project]
name = "tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "gitpython>=3.1.45",
    "google-generativeai>=0.8.5",
    "openai>=2.11.0",
    "requests>=2.32.3",
    "python-dotenv>=1.2.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "pyyaml>=6.0.2",
    "radon>=6.0.1",
]

[project.scripts]
deepreview-cli = "deepreview.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "ruff>=0.6.0",
    "bandit>=1.7.8",
]
